move all execution tests into subfolders
