+++ python_modules/dagster/dagster_tests/cli_tests/command_tests/test_memoized_development_cli.py
@@ -9,7 +9,7 @@ from dagster._cli.job import execute_list_versions_command
 from dagster._core.test_utils import instance_for_test
 from dagster._utils import file_relative_path
 
-from ...execution_tests.memoized_dev_loop_job import op_job
+from ...execution_tests.versioning_tests.memoized_dev_loop_job import op_job
 
 
 class Capturing(list):
